pyproject.toml
[project]
name = "projectvalera-2"
version = "0.1.0"
description = "Вторая настройка проекта с функциями масок для карт и счетов"
authors = [
    {name = "Valera",email = "xoxlov_v83@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "requests (>=2.33.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "projectvalera_2", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
dev = [
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

[tool.mypy]
python_version = "3.14"
warn_unused_configs = true
disallow_untyped_defs = true
warn_return_any = true
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3

[tool.black]
line-length = 119
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | migrations
)/
'''